почти готово, осталось завести пользователей, причесать работу с журналом логов и сделать вид для интервьюера

+++ modules/main/controllers/DivisionController.php
@@ -9,6 +9,7 @@ use yii\data\ActiveDataProvider;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use \yii\web\HttpException;
 
 /**
  * DivisionController implements the CRUD actions for Division model.
@@ -36,22 +37,27 @@ class DivisionController extends Controller
      */
     public function actionIndex()
     {
-        $dataProvider = new ActiveDataProvider([
-            'query' => Division::find(),
-            'pagination' => [
-                'pageSize' => Yii::$app->params['page_size'],
-            ],
-            'sort' => [
-                'attributes' => [
-                    'area' => SORT_ASC,
-                    //'title' => SORT_ASC,
-                ]
-            ],
-        ]);
+        if(Yii::$app->user->can('manager')) {
+            $dataProvider = new ActiveDataProvider([
+                'query' => Division::find(),
+                'pagination' => [
+                    'pageSize' => Yii::$app->params['page_size'],
+                ],
+                'sort' => [
+                    'attributes' => [
+                        'area' => SORT_ASC,
+                        //'title' => SORT_ASC,
+                    ]
+                ],
+            ]);
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
+        }
     }
 
     /**
@@ -61,9 +67,14 @@ class DivisionController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(Yii::$app->user->can('manager')) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
+        }
     }
 
     /**
@@ -73,16 +84,21 @@ class DivisionController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Division();
+        if(Yii::$app->user->can('admin')) {
+            $model = new Division();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            $msg = 'Добавлено новое юрлицо <strong>'. $model->name .'</strong>.';
-            BaseModel::AddEventLog('info',$msg);
-            return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                $msg = 'Добавлено новое юрлицо <strong>' . $model->name . '</strong>.';
+                BaseModel::AddEventLog('info', $msg);
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,
+                ]);
+            }
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
         }
     }
 
@@ -94,16 +110,21 @@ class DivisionController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can('admin')) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            $msg = 'Данные юрлица <strong>'. $model->name .'</strong> были обновлены.';
-            BaseModel::AddEventLog('info',$msg);
-            return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                $msg = 'Данные юрлица <strong>' . $model->name . '</strong> были обновлены.';
+                BaseModel::AddEventLog('info', $msg);
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
         }
     }
 
@@ -115,12 +136,17 @@ class DivisionController extends Controller
      */
     public function actionDelete($id)
     {
-        $row = Division::findOne($id);
-     //   $this->findModel($id)->delete();
-        $row->delete();
-        $msg = 'Данные юрлица <strong>'. $row->name .'</strong> были удалены.';
-        BaseModel::AddEventLog('info',$msg);
-        return $this->redirect(['index']);
+        if(Yii::$app->user->can('admin')) {
+            $row = Division::findOne($id);
+            //   $this->findModel($id)->delete();
+            $row->delete();
+            $msg = 'Данные юрлица <strong>' . $row->name . '</strong> были удалены.';
+            BaseModel::AddEventLog('info', $msg);
+            return $this->redirect(['index']);
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
+        }
     }
 
     /**
