+++ modules/main/controllers/market/FormController.php
@@ -12,6 +12,7 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use app\models\BaseModel;
+use \yii\web\HttpException;
 
 /**
  * FormController implements the CRUD actions for Form model.
@@ -40,16 +41,25 @@ class FormController extends Controller
      */
     public function actionIndex()
     {
-        $dataProvider = new ActiveDataProvider([
-            'query' => Form::find(),
-            'pagination' => [
-                'pageSize' => Yii::$app->params['page_size'],
-            ],
-        ]);
+        if(Yii::$app->user->can('market')) {
+            $dataProvider = new ActiveDataProvider([
+                'query' => Form::find(),
+                'pagination' => [
+                    'pageSize' => Yii::$app->params['page_size'],
+                ],
+            ]);
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+        elseif(Yii::$app->user->can('poll')) {
+            $this->layout = '@app/modules/user/layouts/basic.php';
+            return 'Свой вывод анкет';
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
+        }
     }
 
     /**
@@ -59,50 +69,58 @@ class FormController extends Controller
      */
     public function actionView($id)
     {
-        if(Yii::$app->request->post()){
-            //$request = Yii::$app->request->post();
-            $date = date('Y-m-d'); //текущая дата
-            self::SavePoll($id,$date);
-        }
-        return $this->render('view', [
+        if(Yii::$app->user->can('viewForm')) {
+            if (Yii::$app->request->post()) {
+                //$request = Yii::$app->request->post();
+                $date = date('Y-m-d'); //текущая дата
+                self::SavePoll($id, $date);
+            }
+            return $this->render('view', [
                 'model' => $this->findModel($id),
                 'content' => $this->ViewForm($id),
             ]);
-
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
+        }
     }
 
     public function actionMedia()
     {
-        $id = 7; //Опрос посетителей выставки домов Малоэтажная Страна
+        if(Yii::$app->user->can('guard')) {
+            $id = 7; //Опрос посетителей выставки домов Малоэтажная Страна
 
-        if(\Yii::$app->request->isAjax){
-            $date = $_POST['date'];
-            $kol = $_POST['kolvo'];
-            if($kol > 0){
-                while($kol){
-                    self::SavePoll($id,$date);
-                    $kol--;
+            if (\Yii::$app->request->isAjax) {
+                $date = $_POST['date'];
+                $kol = $_POST['kolvo'];
+                if ($kol > 0) {
+                    while ($kol) {
+                        self::SavePoll($id, $date);
+                        $kol--;
+                    }
                 }
+                return 'OK';
             }
-            return 'OK';
+
+            /*if(Yii::$app->request->post()){
+                //$request = Yii::$app->request->post();
+                $date = $_POST['date'];
+                $kol = $_POST['kolvo'];
+                if($kol > 0){
+                    while($kol){
+                        self::SavePoll($id,$date);
+                        $kol--;
+                    }
+                }
+            }*/
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+                'content' => $this->ViewMedia($id),
+            ]);
         }
-
-        /*if(Yii::$app->request->post()){
-            //$request = Yii::$app->request->post();
-            $date = $_POST['date'];
-            $kol = $_POST['kolvo'];
-            if($kol > 0){
-                while($kol){
-                    self::SavePoll($id,$date);
-                    $kol--;
-                }
-            }
-        }*/
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-            'content' => $this->ViewMedia($id),
-        ]);
-
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
+        }
     }
 
     /**
@@ -112,29 +130,34 @@ class FormController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Form();
+        if(Yii::$app->user->can('market')) {
+            $model = new Form();
 
-        if ($model->load(Yii::$app->request->post()) && $model->validate()) {
-            if($model->is_work){//если хотим установить рабочую анкеты
-                $connection = \Yii::$app->db;
-                $query="update form set is_work=0";
-                $connection->createCommand($query)->execute();
+            if ($model->load(Yii::$app->request->post()) && $model->validate()) {
+                if ($model->is_work) {//если хотим установить рабочую анкеты
+                    $connection = \Yii::$app->db;
+                    $query = "update form set is_work=0";
+                    $connection->createCommand($query)->execute();
+                }
+                $msg = 'Добавлена новая анкета  <strong>' . $model->name . '</strong> пользователем <strong>' . Yii::$app->user->identity->fname . ' ' . Yii::$app->user->identity->lname . '</strong>.';
+                $model->save();
+                BaseModel::AddEventLog('info', $msg);
+                //return $this->redirect(['view', 'id' => $model->id]);
+                return $this->redirect('index');
+            } else {
+                $statsel = array('1' => 'Активная', '0' => 'Не активная');
+                $worksel = array('1' => 'Рабочая', '0' => 'Не рабочая');
+                $qstcount = 0;
+                return $this->render('create', [
+                    'model' => $model,
+                    'statsel' => $statsel,
+                    'worksel' => $worksel,
+                    'qstcount' => $qstcount,
+                ]);
             }
-            $msg = 'Добавлена новая анкета  <strong>'. $model->name .'</strong> пользователем <strong>'.Yii::$app->user->identity->fname .' '.Yii::$app->user->identity->lname.'</strong>.';
-            $model->save();
-            BaseModel::AddEventLog('info',$msg);
-            //return $this->redirect(['view', 'id' => $model->id]);
-            return $this->redirect('index');
-        } else {
-            $statsel = array ('1' => 'Активная','0' => 'Не активная');
-            $worksel = array ('1' => 'Рабочая','0' => 'Не рабочая');
-            $qstcount = 0;
-            return $this->render('create', [
-                'model' => $model,
-                'statsel' => $statsel,
-                'worksel' => $worksel,
-                'qstcount' => $qstcount,
-            ]);
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
         }
     }
 
@@ -146,29 +169,34 @@ class FormController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can('market')) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->validate()) {
-            if($model->is_work){//если хотим установить рабочую анкеты
-                $connection = \Yii::$app->db;
-                $query="update form set is_work=0";
-                $connection->createCommand($query)->execute();
+            if ($model->load(Yii::$app->request->post()) && $model->validate()) {
+                if ($model->is_work) {//если хотим установить рабочую анкеты
+                    $connection = \Yii::$app->db;
+                    $query = "update form set is_work=0";
+                    $connection->createCommand($query)->execute();
+                }
+                $msg = 'Данные анкеты  <strong>' . $model->name . '</strong> были обновлены пользователем <strong>' . Yii::$app->user->identity->fname . ' ' . Yii::$app->user->identity->lname . '</strong>.';
+                $model->save();
+                BaseModel::AddEventLog('info', $msg);
+                //return $this->redirect(['view', 'id' => $model->id]);
+                return $this->redirect('index');
+            } else {
+                $statsel = array('1' => 'Активная', '0' => 'Не активная');
+                $worksel = array('1' => 'Рабочая', '0' => 'Не рабочая');
+                $qstcount = Questions::find()->where(['=', 'form_id', $id])->count();
+                return $this->render('update', [
+                    'model' => $model,
+                    'statsel' => $statsel,
+                    'worksel' => $worksel,
+                    'qstcount' => $qstcount,
+                ]);
             }
-            $msg = 'Данные анкеты  <strong>'. $model->name .'</strong> были обновлены пользователем <strong>'.Yii::$app->user->identity->fname .' '.Yii::$app->user->identity->lname.'</strong>.';
-            $model->save();
-            BaseModel::AddEventLog('info',$msg);
-            //return $this->redirect(['view', 'id' => $model->id]);
-            return $this->redirect('index');
-        } else {
-            $statsel = array ('1' => 'Активная','0' => 'Не активная');
-            $worksel = array ('1' => 'Рабочая','0' => 'Не рабочая');
-            $qstcount = Questions::find()->where(['=','form_id',$id])->count();
-            return $this->render('update', [
-                'model' => $model,
-                'statsel' => $statsel,
-                'worksel' => $worksel,
-                'qstcount' => $qstcount,
-            ]);
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
         }
     }
 
@@ -180,17 +208,22 @@ class FormController extends Controller
      */
     public function actionDelete($id)
     {
-        //есть ли вопросы
-        $qst = Questions::find()->where(['=','form_id',$id])->count();
-        if($qst)
-            Yii::$app->session->setFlash('error', 'У анкеты есть не удаленные вопросы. Удаление не возможно!');
-        else {
-            $msg = 'Анкета  <strong>'. $this->findModel($id)->name .'</strong> была удалена пользователем <strong>'.Yii::$app->user->identity->fname .' '.Yii::$app->user->identity->lname.'</strong>.';
-            $this->findModel($id)->delete();
-            BaseModel::AddEventLog('info',$msg);
-        }
+        if(Yii::$app->user->can('market')) {
+            //есть ли вопросы
+            $qst = Questions::find()->where(['=', 'form_id', $id])->count();
+            if ($qst)
+                Yii::$app->session->setFlash('error', 'У анкеты есть не удаленные вопросы. Удаление не возможно!');
+            else {
+                $msg = 'Анкета  <strong>' . $this->findModel($id)->name . '</strong> была удалена пользователем <strong>' . Yii::$app->user->identity->fname . ' ' . Yii::$app->user->identity->lname . '</strong>.';
+                $this->findModel($id)->delete();
+                BaseModel::AddEventLog('info', $msg);
+            }
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        else{
+            throw new HttpException(404 ,'Доступ запрещен');
+        }
     }
 
     /**
